refactor: CartItem 컴포넌트 속 CartItemInfo, Counter 컴포넌트 .tsx파일 분리 - Issue #5
